继续清除裸时间。

+++ src/System/SystemConfiguration.h
@@ -88,7 +88,7 @@ public:
 	}
 
 	// World market
-	long GetWorldMarketResettingTime() const noexcept { return m_lMarketResettingTime; }
+	unsigned GetWorldMarketResettingTime() const noexcept { return m_lMarketResettingTime; }
 	[[nodiscard]] string GetFinnhubToken() noexcept { return m_strFinnhubToken; }
 	[[nodiscard]] string GetQuandlToken() noexcept { return m_strQuandlToken; }
 	[[nodiscard]] auto GetWorldMarketFinnhubInquiryTime() const noexcept { return m_worldMarketFinnhubInquiryTime; } // 单位为毫秒
@@ -352,7 +352,7 @@ protected:
 	chrono::milliseconds m_chinaMarketRTDataInquiryTime{ 250 }; // 中国市场实时数据查询间隔时间,单位为毫秒
 
 	// World Market
-	long m_lMarketResettingTime{ 170000 }; // 默认市场重置时间为170000
+	unsigned m_lMarketResettingTime{ 170000 }; // 默认市场重置时间为170000
 	string m_strFinnhubToken{ "" }; // 令牌
 	string m_strQuandlToken{ "" }; // 令牌
 	bool m_bFinnhubAccountFeePaid{ true }; // 付费账户或者免费账户
